pyproject.toml
[tool.poetry]
name = "FastAPI course"
version = "0.2.0"
description = "Course to learn FastAPI and make my own API RestFul, adding Github actions"
authors = ["Jkrox <jkrox@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.98.0"
uvicorn = "^0.22.0"
pydantic = "^1.10.9"
pytest = "^7.4.0"
black = "^23.3.0"
pyjwt = "^2.7.0"
python-dotenv = "^1"
sqlalchemy = "^2.0.19"
pre-commit = "^3.3.3"
flake8 = "^6.1.0"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# pyproject.toml
[tool.black]
line-length = 79  # Ajusta la longitud máxima de línea según tus preferencias
